+++ cmd/som/cli/plot.go
@@ -15,7 +15,17 @@ func plotCommand() *cobra.Command {
 	command := &cobra.Command{
 		Use:   "plot",
 		Short: "Plots visualizations for an SOM in various ways. See sub-commands.",
-		Long:  `Plots visualizations for an SOM in various ways. See sub-commands.`,
+		Long: `Plots visualizations for an SOM in various ways. See sub-commands.
+
+All sub-commands take an SOM YAML file and an output PNG file
+as positional arguments.
+
+Please note that the built-in visualizations are not intended for
+publication-quality output. Instead, they serve as quick tools for
+inspecting training and prediction results. For high-quality visualizations,
+we recommend exporting the SOM and other results to CSV files.
+You can then use dedicated visualization libraries in languages such as
+Python or R to create more refined and customized graphics.`,
 		Run: func(cmd *cobra.Command, args []string) {
 			_ = cmd.Help()
 		},